Small updates regarding libsecp256k1
secp256k1 is the name of the curve and for historic reasons, also the repository name of
the Bitcoin crypto library libsecp256k1. It implements ECDSA and Schnorr signatures on
the secp256k1 curve. Multisigning support is a work in progress currently.

secp256k1-zkp is a fork of libsecp256k1 which adds Borromean Schnorr ring signatures,
Petersen commitments, and range proofs. It's specifically used for implementing the
Confidential Transactions technology used in Elements Alpha.