Several edits to main text + commentary re 2nd paragraph.
Following the 2nd paragraph of the Overview...I have this commentary. 

I really don't know that this is TRUE - I don't believe there is research to prove this...networks by their nature lead to hyperconnected nodes HAVEING a lot of power. Look at all the work done about small scale networks and power laws and understanding what happened to the internet - where we are today is ALL power law networks and the power imbalances that creates - these type of thigns WILL happen in whatever new network thing we think we are creating. It is a falicy to think otherwise. This is why I pulled ALL this text from the version I was working on yesterday - it is based on a bunch of assumptions that are NOT realistic. They are techno-utopian mumbo jumob,  I also think it is dangerous to broadly generalize about how democracies work at balancing power. I think they are failing miserably - having been bought by the monied and powerful - ironically enough...those large tech companies at heart of network effects we are trying to address/could end up crating even within new DPKI.