Serious Questions about 2nd paragraph
In the either pad edition I did a bunch of editing and in earlier versions and took out the 2nd paragraph and added some other context (the other context is still in the above version but the 2nd paragraph got added back in. I don't think it makes rhetorical logical sense.  I don't think the statements about democracy - networks and power are proven by research (At least none that I have heard of). Networks by their nature lead to hyperconnected nodes HAVEING a lot of power. Look at all the work done about small scale networks and power laws and understanding what happened to the internet - where we are today is ALL power law networks and the power imbalances that creates - these type of thigns WILL happen in whatever new network thing we think we are creating. It is a falicy to think otherwise. This is why I pulled ALL this text from the version I was working on yesterday - it is based on a bunch of assumptions that are NOT realistic. They are techno-utopian mumbo jumob.  I also think pointing to what seems like an assumption of american democracy that is has in effect been bought by special interests - those with money.