removed ending it is far to broad and presumptions a statement.
Just who do we think is in this movement and what does it encompass? Are you seeking to build a sytem that has the "trust" and confidence of most of the world's population becuase it was not in the room 2 weeks ago.  This sort of presumption that these types of "we are the world" and the people will join us if we just build it frames are dangerous. 

What is your evidence that it is expanding and growing.  I am still very keen to have "us" the people who gathered NOT come out to the world and "say" this is the name for what DPKI beyond doing e-mail ... but for real things is "called" - it is packing TO MUCH into a word that doesn't for normal people mean what is identifiers + crypto and the ability to strong authentication and encrypted messages... maybe some new functionality - but to say "trust" for what is cryptographically good but not socially meaningful is I believe the wrong approach to success.