Critical point missed - Abstract representation of VALUE
The point with Hernando De Soto's work that is critical is that he highlights what the system of deeds and other similar systems do in the WEST they create ways to abstractly represent things exist the real world  - in ways that those abstractions take on their own life/power. THIS is critical not that some how the deeds have "force of law" it is isn't that...it is that people believe in the deeds an the deeds - the abstract representation of the value of the land that they own can be used to do other things. This translates into people and their identities - they ABSTRACT representation of myself in the form of government issued paperwork (pre digital) and then in the digital world of digital identifiers - these can be digital identifiers that I personally own and control they do not have to drive their power "From" a government issuing them to me.  THIS is the point.  
I already made these edits before and they were not accepted and you didn't discuss them with me. If the changes I made are not clear or you don't "get" what i am trying to say then please contact me. Do not simply not accept the point because it was in the original text that I wrote and it is the critical point I was seeking to make in the original writing.