Create Principle of Relativity for WoT
By way of an analogy to Einstein’s theory of relativity, I propose a principle of relativity for reputation in the context of the web of trust (WoT). This principle states that reputation can only be properly understood relative to a specific end-user. In other words, reputation is in the eye of the beholder. By this principle, we should shift focus from global reputation (e.g. Bob’s trust score in absolute terms) to local reputation (e.g., Alice’s evaluation of Bob’s trustworthiness). Indeed, any attempt to construct WoT in a manner that enables the calculation of global reputation necessarily exacts a cost to privacy and in so doing will be detrimental to the overall utility of the WoT.